Lab2/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
Pillow
tqdm
